fix build error

+++ localization/ndt_scan_matcher/src/ndt_scan_matcher_core.cpp
@@ -278,7 +278,7 @@ void NDTScanMatcher::callback_sensor_points(
     callback_sensor_points_main(sensor_points_msg_in_sensor_frame);
 
   // check skipping_publish_num
-  static size_t skipping_publish_num = 0;
+  static int64_t skipping_publish_num = 0;
   skipping_publish_num =
     ((is_succeed_scan_matching || !is_activated_) ? 0 : (skipping_publish_num + 1));
   diagnostics_scan_points_->addKeyValue("skipping_publish_num", skipping_publish_num);
